Using Arrays in Java
I wrote a program for taking marks input for n number of students from the user.

The program first asks to enter the number of students and then it creates an array of that size and stores n number of names in that array.

Later, the program asks user to input marks for each student and at last it ends by printing Name and Marks together for the total no. of students that were there.